Initial Commit
This is my initial commit for this project.  I am setting up the files 
that I will use for this project.  I am taking the template provided in 
the Udacity project and I am renaming them to make them my own.  I will 
edit the template files provided and show my commits as I proceed.  This 
should be fun!