Changing Project
I decided to go with a different file to work with and explore.  Instead 
of using the two Bay Wheels Trip Data that are not structured the same, 
I am going with the Prosper Loan Data.  I don't know if it will be as 
fun, but it may be more simple.  Therefore, I am removing the two 
baywheels csv files and creating a new Jupyter Notebook for this new 
data set.